% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_multiple.R
\name{compute_ts_ci_area}
\alias{compute_ts_ci_area}
\title{Compute time-series confidence intervals area}
\usage{
compute_ts_ci_area(x, group, alpha = 0.05)
}
\arguments{
\item{x}{Numeric vector.}

\item{group}{Grouping vector of the same length as \code{x}.}

\item{alpha}{Numeric scalar. Size of the two-sided interval tail.}
}
\value{
A list with two data frames: \code{area} (for polygons) and \code{line}
(for the median time-series).
}
\description{
This function computes the area for confidence intervals of time-series
data, which can be used for visualizing uncertainty in time-series plots.
It takes a numeric vector \code{x}, a grouping vector \code{group} that indicates
which time point each value in \code{x} corresponds to, and an \code{alpha} level for
the confidence interval. The function returns a list containing two data
frames: one for the area of the confidence interval (suitable for polygon
plotting) and one for the median line of the time-series.
}
\examples{
# Simulating random walks
set.seed(123)
dat <- lapply(1:200, \(i) {
  data.frame(
    x = cumsum(rnorm(10)),
    time = 1:10
  )
}) |> do.call(what = "rbind")
ans <- compute_ts_ci_area(dat$x, dat$time)
plot(ans$line, type = "b", ylim = c(-6, 6))
polygon(x = ans$area$x, y = ans$area$y)
}
\concept{misc-functions}
