% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{epiworld-hospitalizations}
\alias{epiworld-hospitalizations}
\alias{get_hospitalizations}
\title{Hospitalizations by tool}
\usage{
get_hospitalizations(x)
}
\arguments{
\item{x}{An object of class \code{\link{epiworld_sir}}, \code{\link{epiworld_seir}}, etc.
(any model).}
}
\value{
\itemize{
\item The function \code{get_hospitalizations} returns a data.frame with five columns:
\code{date}, \code{virus_id}, \code{tool_id}, \code{hospitalizations}, and \code{weight}. The \code{weight}
column is used to keep track of individuals having multiple tools. For
example, if an agent has two tools (vaccination and mask-wearing), then it
will show up twice under count, but with weights 0.5 for each count. Models
with no hospitalization tracking will return the same data.frame with no
rows.
}
}
\description{
Weighted hospitalization tracking when agents have multiple tools.
}
\examples{
# See model documentation for examples with hospitalization tracking
}
\seealso{
Other Summaries:
\code{\link{epiworld-summaries}}
}
\concept{Summaries}
\concept{model-utility-functions}
