% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiworldR-package.R
\name{epiworld_cpp_version}
\alias{epiworld_cpp_version}
\title{Version of the epiworld C++ code}
\usage{
epiworld_cpp_version()
}
\value{
A character string representing the version of the C++ library.
}
\description{
Returns the version of the C++ library epiworld. The code
is hosted on GitHub at \url{https://github.com/UofUEpiBio/epiworld}.
}
\examples{
epiworld_cpp_version()

}
\keyword{internal}
