% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiworldR-package.R
\docType{package}
\name{epiworldRfaster-package}
\alias{epiworldRfaster}
\alias{epiworldRfaster-package}
\title{epiworldRfaster}
\description{
A flexible framework for Agent-Based Models (ABM), the 'epiworldR' package provides methods for prototyping disease outbreaks and transmission models using a 'C++' backend, making it very fast. It supports multiple epidemiological models, including the Susceptible-Infected-Susceptible (SIS), Susceptible-Infected-Removed (SIR), Susceptible-Exposed-Infected-Removed (SEIR), and others, involving arbitrary mitigation policies and multiple-disease models. Users can specify infectiousness/susceptibility rates as a function of agents' features, providing great complexity for the model dynamics. Furthermore, 'epiworldR' is ideal for simulation studies featuring large populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofUEpiBio/epiworldR}
  \item \url{https://uofuepibio.github.io/epiworldR/}
  \item \url{https://uofuepibio.github.io/epiworldR-workshop/}
  \item Report bugs at \url{https://github.com/UofUEpiBio/epiworldR/issues}
}

}
\author{
\strong{Maintainer}: Derek Meyer \email{derekmeyer37@gmail.com} (\href{https://orcid.org/0009-0005-1350-6988}{ORCID})

Authors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})
}

Other contributors:
\itemize{
  \item Susan Holmes (\href{https://orcid.org/0000-0002-2208-8168}{ORCID}) (JOSS reviewer) [reviewer]
  \item Abinash Satapathy (\href{https://orcid.org/0000-0002-2955-2744}{ORCID}) (JOSS reviewer) [reviewer]
  \item Carinogurjao [reviewer]
  \item Centers for Disease Control and Prevention (Award number 1U01CK000585; 75D30121F00003) [funder]
}

}
