% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiworldR-package.R
\name{first_function}
\alias{first_function}
\alias{second_function}
\title{Epi Functions}
\usage{
first_function(x, y, z)

second_function(x)
}
\arguments{
\item{x}{This is a number}

\item{y}{This is another number}

\item{z}{This is the last number}
}
\value{
The sum of \code{x}, \code{y}, and \code{z}.
}
\description{
Epi Functions
}
\examples{
first_function(1, 1, 3)
first_function(13, 2, 400)
}
