% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents-methods.R
\name{get_agents}
\alias{get_agents}
\alias{[.epiworld_agents}
\alias{print.epiworld_agent}
\alias{print.epiworld_agents}
\alias{get_state}
\title{Agents in epiworldR}
\usage{
get_agents(model)

\method{[}{epiworld_agents}(x, i)

\method{print}{epiworld_agent}(x, compressed = FALSE, ...)

\method{print}{epiworld_agents}(x, compressed = TRUE, max_print = 10, ...)

get_state(x)
}
\arguments{
\item{model}{An object of class \link{epiworld_model}.}

\item{x}{An object of class \link{epiworld_agents}}

\item{i}{Index (id) of the agent (from 0 to \code{n-1})}

\item{compressed}{Logical scalar. When FALSE, it prints detailed information
about the agent.}
}
\description{
Agents in epiworldR
}
\seealso{
agents
}
