% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-actions.R
\name{global-actions}
\alias{global-actions}
\alias{globalaction_tool}
\alias{globalaction_tool_logit}
\alias{add_global_action}
\title{Global Actions}
\usage{
globalaction_tool(tool, prob)

globalaction_tool_logit(tool, vars, coefs)

add_global_action(model, action, date = -99)
}
\arguments{
\item{tool}{A tool.}

\item{prob}{Numeric. A probability.}

\item{action}{A global action.}

\item{date}{Integer. The date at which the action is executed (see details).}
}
\description{
Global actions are functions that are executed at each time step of the
simulation. They are useful for implementing interventions, such as
vaccination, isolation, and social distancing by means of tools.
}
\details{
The function \code{globalaction_tool_logit} allows to specify a logistic
regression model for the probability of using a tool. The model is specified
by the vector of coefficients \code{coefs} and the vector of variables \code{vars}.
\code{vars} is an integer vector indicating the position of the variables in the
model.

The function \code{add_global_action} adds a global action to a model.
The model checks for actions to be executed at each time step. If the added
action matches the current time step, the action is executed. When \code{date} is
negative, the action is executed at each time step. When \code{date} is positive,
the action is executed at the specified time step.
}
\examples{
# Simple model
model_sirconn <- ModelSIRCONN(
  name                = "COVID-19",
  n                   = 10000,
  prevalence          = 0.01,
  contact_rate        = 5,
  prob_transmission   = 0.4,
  prob_recovery       = 0.95
)

# Creating a tool
epitool <- tool(
  name = "Vaccine",
  susceptibility_reduction = .9,
  transmission_reduction = .5,
  recovery_enhancer = .5, 
  death_reduction = .9
)


# Adding a global action
vaccine_day_20 <- globalaction_tool(epitool, .2)
add_global_action(model_sirconn, vaccine_day_20, 20)

# Running and printing
run(model_sirconn, ndays = 100, seed = 1912)
model_sirconn
plot(model_sirconn)
}
\seealso{
epiworld-model
}
