% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measles-package.R
\docType{package}
\name{measles-package}
\alias{measles}
\alias{measles-package}
\title{measles: Measles Epidemiological Models}
\description{
A specialized collection of measles epidemiological models built on the
epiworldR framework. This package provides models for studying measles
transmission dynamics, including school settings with quarantine and
isolation policies, mixing models with population groups, and risk-based
quarantine strategies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofUEpiBio/measles}
  \item Report bugs at \url{https://github.com/UofUEpiBio/measles/issues}
}

}
\author{
\strong{Maintainer}: George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})

Other contributors:
\itemize{
  \item Centers for Disease Control and Prevention (Award number 1U01CK000585; 75D30121F00003) [funder]
}

}
\keyword{internal}
