% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIR.R
\name{ModelSEIR}
\alias{ModelSEIR}
\alias{epiworld_seir}
\alias{init.epiworld_seir}
\alias{print.epiworld_seir}
\alias{agents_smallworld.epiworld_seir}
\alias{run.epiworld_seir}
\alias{plot.epiworld_seir}
\title{Susceptible Infected Susceptible model (SEIR)}
\usage{
ModelSEIR(name, prevalence, infectiousness, incubation_days, recovery)

\method{init}{epiworld_seir}(m, days, seed)

\method{print}{epiworld_seir}(x, ...)

\method{agents_smallworld}{epiworld_seir}(m, n, k, d, p)

\method{run}{epiworld_seir}(m)

\method{plot}{epiworld_seir}(x, ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{infectiousness}{Numeric scalar between 0 and 1. Virus's rate of
infection.}

\item{incubation_days}{Numeric scalar greater than 0. Average number of
incubation days.}

\item{recovery}{Numeric scalar between 0 and 1. Rate of recovery from virus.}

\item{m}{to be documented}

\item{days}{to be documented}

\item{seed}{to be documented}

\item{x}{Object of class SIS.}

\item{...}{Currently ignore.}

\item{n}{to be documented}

\item{k}{to be documented}

\item{d}{to be documented}

\item{p}{to be documented}
}
\description{
Susceptible Infected Susceptible model (SEIR)
}
\examples{
model_seir <- ModelSEIR(name = "COVID-19", prevalence = 0.01, 
infectiousness = 0.9, recovery = 0.1, incubation_days = 4)

# Adding a small world population
agents_smallworld(
  model_seir,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Initializing
init(model_seir, days = 100, seed = 1912)
# Running and printing
run(model_seir)
model_seir
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
