% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIRCONN.R
\name{ModelSEIRCONN}
\alias{ModelSEIRCONN}
\alias{epiworld_seirconn}
\alias{init.epiworld_seirconn}
\alias{print.epiworld_seirconn}
\alias{run.epiworld_seirconn}
\title{Susceptible Exposed Infected Removed model (SEIR connected)}
\usage{
ModelSEIRCONN(
  name,
  n,
  prevalence,
  reproductive_number,
  prob_transmission,
  incubation_days,
  prob_recovery
)

\method{init}{epiworld_seirconn}(m, days, seed)

\method{print}{epiworld_seirconn}(x, ...)

\method{run}{epiworld_seirconn}(m)
}
\arguments{
\item{name}{Name of the virus}

\item{n}{Integer greater than zero. Population size.}

\item{prevalence}{Initial proportion of individuals with the virus}

\item{reproductive_number}{Numeric scalar. Reproductive number.}

\item{prob_transmission}{Numeric scalar between 0 and 1. Probability of transmission.}

\item{incubation_days}{Numeric scalar greater than 0. Average number of incubation days.}

\item{prob_recovery}{a Numeric scalar between 0 and 1. Probability of recovery.}

\item{m}{to be documented}

\item{days}{to be documented}

\item{seed}{to be documented}

\item{x}{to be documented}

\item{...}{to be documented}
}
\description{
The SEIR connected model implements a model where all agents are connected.
This is equivalent to a comparmental model.
}
\seealso{
Other Models: 
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
