% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIRD.R
\name{ModelSEIRD}
\alias{ModelSEIRD}
\alias{epiworld_seird}
\alias{plot.epiworld_seird}
\title{Susceptible-Exposed-Infected-Recovered-Deceased model (SEIRD)}
\usage{
ModelSEIRD(
  name,
  prevalence,
  transmission_rate,
  incubation_days,
  recovery_rate,
  death_rate
)

\method{plot}{epiworld_seird}(x, main = get_name(x), ...)
}
\arguments{
\item{name}{String. Name of the virus.}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{transmission_rate}{Numeric scalar between 0 and 1. Virus's rate of
infection.}

\item{incubation_days}{Numeric scalar greater than 0. Average number of
incubation days.}

\item{recovery_rate}{Numeric scalar between 0 and 1. Rate of recovery_rate from virus.}

\item{death_rate}{Numeric scalar between 0 and 1. Rate of death from virus.}

\item{x}{Object of class SEIRD.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\value{
\itemize{
\item The \code{ModelSEIRD}function returns a model of class \link{epiworld_model}.
}

The \code{plot} function returns a plot of the SEIRD model of class
\link{epiworld_model}.
}
\description{
Susceptible-Exposed-Infected-Recovered-Deceased model (SEIRD)
}
\examples{
model_seird <- ModelSEIRD(name = "COVID-19", prevalence = 0.01, 
transmission_rate = 0.9, recovery_rate = 0.1, incubation_days = 4, 
death_rate = 0.001)

# Adding a small world population
agents_smallworld(
  model_seird,
  n = 10000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Running and printing
run(model_seird, ndays = 100, seed = 1912)
model_seird

plot(model_seird, main = "SEIRD Model")
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelDiffNet}()},
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIRDCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIRDCONN}()},
\code{\link{ModelSIRD}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSISD}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
