% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIRMixing.R
\name{ModelSEIRMixing}
\alias{ModelSEIRMixing}
\alias{epiworld_seirmixing}
\alias{plot.epiworld_seirmixing}
\title{Susceptible Exposed Infected Removed model (SEIR) with mixing}
\usage{
ModelSEIRMixing(
  name,
  n,
  prevalence,
  contact_rate,
  transmission_rate,
  incubation_days,
  recovery_rate,
  contact_matrix
)

\method{plot}{epiworld_seirmixing}(x, main = get_name(x), ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{n}{Number of individuals in the population.}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{contact_rate}{Numeric scalar. Average number of contacts per step.}

\item{transmission_rate}{Numeric scalar between 0 and 1. Probability of
transmission.}

\item{incubation_days}{Numeric scalar. Average number of days in the
incubation period.}

\item{recovery_rate}{Numeric scalar between 0 and 1. Probability of recovery.}

\item{contact_matrix}{Matrix of contact rates between individuals.}

\item{x}{Object of class SIRCONN.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\value{
\itemize{
\item The \code{ModelSEIRMixing}function returns a model of class \link{epiworld_model}.
}

The \code{plot} function returns a plot of the SEIRMixing model of class
\link{epiworld_model}.
}
\description{
Susceptible Exposed Infected Removed model (SEIR) with mixing
}
\details{
The \code{contact_matrix} is a matrix of contact rates between entities. The
matrix should be of size \verb{n x n}, where \code{n} is the number of entities.
This is a row-stochastic matrix, i.e., the sum of each row should be 1.

The \link{initial_states} function allows the user to set the initial state of the
model. In particular, the user can specify how many of the non-infected
agents have been removed at the beginning of the simulation.
}
\examples{
# TBD
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelDiffNet}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIRD}()},
\code{\link{ModelSEIRDCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIRD}()},
\code{\link{ModelSIRDCONN}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIRMixing}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSISD}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
