% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIR.R
\name{ModelSIR}
\alias{ModelSIR}
\alias{epiworld_sir}
\alias{init.epiworld_sir}
\alias{print.epiworld_sir}
\alias{agents_smallworld.epiworld_sir}
\alias{run.epiworld_sir}
\alias{plot.epiworld_sir}
\title{SIR model}
\usage{
ModelSIR(name, prevalence, infectiousness, recovery)

\method{init}{epiworld_sir}(m, days, seed)

\method{print}{epiworld_sir}(x, ...)

\method{agents_smallworld}{epiworld_sir}(m, n, k, d, p)

\method{run}{epiworld_sir}(m)

\method{plot}{epiworld_sir}(x, ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{infectiousness}{Numeric scalar between 0 and 1. Virus's rate of
infection.}

\item{recovery}{Numeric scalar between 0 and 1. Rate of recovery from virus.}

\item{x}{Object of class SIS.}

\item{...}{Currently ignore.}

\item{n}{Number of individuals in the population.}

\item{k}{Number of ties in the small world network.}

\item{d}{Logical scalar. Whether the graph is directed or not.}

\item{p}{Probability of rewiring.}
}
\description{
SIR model
}
\examples{
model_sir <- ModelSIR(name = "COVID-19", prevalence = 0.01, 
                      infectiousness = 0.9, recovery = 0.1)

# Adding a small world population
agents_smallworld(
  model_sir,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Initializing
init(model_sir, days = 100, seed = 1912)
# Running and printing
run(model_sir)
model_sir
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
