% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIRCONN.R
\name{ModelSIRCONN}
\alias{ModelSIRCONN}
\alias{epiworld_sirconn}
\alias{init.epiworld_sirconn}
\alias{print.epiworld_sirconn}
\alias{run.epiworld_sirconn}
\alias{plot.epiworld_sirconn}
\title{Susceptible Infected Removed model (SIR connected)}
\usage{
ModelSIRCONN(
  name,
  n,
  prevalence,
  reproductive_number,
  prob_transmission,
  prob_recovery
)

\method{init}{epiworld_sirconn}(m, days, seed)

\method{print}{epiworld_sirconn}(x, ...)

\method{run}{epiworld_sirconn}(m)

\method{plot}{epiworld_sirconn}(x, ...)
}
\arguments{
\item{name}{Name of the virus}

\item{n, m, days, seed, x, ...}{to be documented}

\item{prevalence}{a number}

\item{reproductive_number}{a number}

\item{prob_transmission}{a number}

\item{prob_recovery}{a number}
}
\description{
Susceptible Infected Removed model (SIR connected)
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
