% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIRLogit.R
\name{ModelSIRLogit}
\alias{ModelSIRLogit}
\title{SIR Logistic model}
\usage{
ModelSIRLogit(
  vname,
  data,
  coefs_infect,
  coefs_recover,
  coef_infect_cols,
  coef_recover_cols,
  prob_infection,
  prob_recovery,
  prevalence
)
}
\arguments{
\item{vname}{}

\item{data}{}

\item{coefs_infect}{}

\item{coefs_recover}{}

\item{coef_infect_cols}{}

\item{coef_recover_cols}{}

\item{prevalence}{}
}
\description{
SIR Logistic model
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
