% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIS.R
\name{ModelSIS}
\alias{ModelSIS}
\alias{epiworld_sis}
\alias{init.epiworld_sis}
\alias{print.epiworld_sis}
\alias{agents_smallworld.epiworld_sis}
\alias{run.epiworld_sis}
\title{Susceptible Infected Susceptible model (SIS)}
\usage{
ModelSIS(name, prevalence, infectiousness, recovery)

\method{init}{epiworld_sis}(m, days, seed)

\method{print}{epiworld_sis}(x, ...)

\method{agents_smallworld}{epiworld_sis}(m, n, k, d, p)

\method{run}{epiworld_sis}(m)
}
\arguments{
\item{name}{Name of the virus}

\item{prevalence}{a number}

\item{infectiousness}{a number}

\item{recovery}{a number}

\item{m, days, seed, x, ..., n, k, d, p}{to be documented}
}
\description{
Susceptible Infected Susceptible model (SIS)
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
