% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network-models}
\alias{network-models}
\alias{agents_sbm}
\alias{agents_sbm.epiworld_model}
\title{Network models in epiworld}
\usage{
agents_sbm(model, block_sizes, mixing_matrix)

\method{agents_sbm}{epiworld_model}(model, block_sizes, mixing_matrix)
}
\arguments{
\item{model}{Model object of class \link{epiworld_model}.}

\item{block_sizes}{Integer vector describing the size of each block in the SBM.}

\item{mixing_matrix}{Numeric matrix describing the mixing between blocks in the SBM.}
}
\value{
\itemize{
\item The function \code{agents_sbm} generates an stochastic block model (SBM) network.
}
}
\description{
Network models in epiworld
}
\examples{
# Initializing SIR model with SBM network
sir <- ModelSIR(name = "COVID-19", prevalence = 0.01, transmission_rate = 0.9,
  recovery_rate = 0.1)
agents_sbm(
  sir,
  block_sizes = c(500, 500),
  mixing_matrix = matrix(c(0.9, 0.1, 0.1, 0.9), nrow = 2)
)
)
run(sir, ndays = 100, seed = 1912)
sir
}
