% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epi.R
\name{plot.epiworld_model}
\alias{plot.epiworld_model}
\title{Plot epidemic curves}
\usage{
\method{plot}{epiworld_model}(x, auto_trunc = FALSE, main = get_name(x), ...)
}
\arguments{
\item{x}{An object of class \code{epiworld_model} or \code{epiworld_hist_virus}
or \code{epiworld_hist}.}

\item{auto_trunc}{A logical value indicating whether to automatically
truncate the plot (y-axis) (see details).}

\item{main}{The main title for the plot.}

\item{counts_scale}{The scale for the counts axis.}
}
\value{
A plot of the epidemic curves for the specified model or history object.
}
\description{
Plot epidemic curves
}
\details{
If \code{auto_trunc} is set to \code{TRUE}, the function will automatically truncate
the plot based on the maximum date when the counts stop significantly
changing by state. The benchmark value for determining significant change
is set to 0.5\% of the range of counts.
}
\examples{
# Building and Initializing SEIR Model
sir <- ModelSIR(
  name = "COVID-19", prevalence = 0.01,
  transmission_rate = 0.9,
  recovery_rate = 0.1
)

# Adding a Small world population
agents_smallworld(
  sir,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
)
# Running and printing
run(sir, ndays = 100, seed = 1912)
plot(sir, main = "SIR Model")
}
\concept{plot-epidemic-curves}
