% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_multiple.R
\name{plot.epiworld_multiple_save}
\alias{plot.epiworld_multiple_save}
\title{Visualize results from \code{run_multiple()}}
\usage{
\method{plot}{epiworld_multiple_save}(x, y = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{epiworld_run_multiple}}

\item{y}{Ignored.}

\item{...}{Further arguments passed to the method.}
}
\description{
Visualize results from \code{run_multiple()}
}
\details{
In the case of \code{transition} objects, the function will collapse the
transition matrix and call \code{\link[=draw_mermaid_from_matrix]{draw_mermaid_from_matrix()}} followed by
\code{plot()}.
}
\concept{plot-epidemic-curves}
