% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_multiple.R
\name{plot_epiworld_multiple_save_transition}
\alias{plot_epiworld_multiple_save_transition}
\title{Plot transition matrices from \code{run_multiple_get_results()}}
\usage{
plot_epiworld_multiple_save_transition(x, ...)
}
\arguments{
\item{x}{A data frame in transition format, usually
\code{run_multiple_get_results(model)$transition}.}

\item{...}{Further arguments passed to \code{\link[=draw_mermaid_from_matrix]{draw_mermaid_from_matrix()}}.}
}
\description{
Plot transition matrices from \code{run_multiple_get_results()}
}
