% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_run_multiple.R
\name{plot_multiple_transition}
\alias{plot_multiple_transition}
\title{Plot transition matrices from \code{run_multiple_get_results()}}
\usage{
plot_multiple_transition(x, ...)
}
\arguments{
\item{x}{A data frame in transition format, usually
\code{run_multiple_get_results(model)$transition}.}

\item{...}{Further arguments passed to \code{\link[=draw_mermaid_from_matrix]{draw_mermaid_from_matrix()}}.}
}
\description{
This function takes the transition data frame from
\code{run_multiple_get_results()}, particularly, the \code{transition} component,
and visualizes it as a transition matrix using \code{draw_mermaid_from_matrix()}.
}
\details{
The function is automatically called when plotting an object of class
\code{epiworld_multiple_save}. It constructs a transition matrix from the
provided data frame, normalizes the rows to represent probabilities, and
then visualizes it using a mermaid diagram. The resulting plot illustrates
the transitions between different states in the model, with the thickness of
the arrows corresponding to the transition.
}
\concept{epiworld-model-diagrams}
