% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{tool}
\alias{tool}
\alias{add_tool}
\alias{add_tool_n}
\alias{rm_tool}
\title{Tools in epiworld}
\usage{
tool(
  name,
  susceptibility_reduction,
  transmission_reduction,
  recovery_enhancer,
  death_reduction
)

add_tool(m, t, prevalence)

add_tool_n(m, t, prevalence)

rm_tool(m, tool_pos)

rm_tool(m, tool_pos)
}
\arguments{
\item{name}{Name of the tool}

\item{susceptibility_reduction}{Numeric. Proportion it reduces susceptibility.}

\item{transmission_reduction}{Numeric. Proportion it reduces transmission.}

\item{recovery_enhancer}{Numeric. Proportion it improves recovery.}

\item{death_reduction}{Numeric. Proportion it reduces probability of death.e}

\item{m}{Model}

\item{t}{An object of class \code{epiworld_tool}}

\item{prevalence}{In the case of \code{add_tool}, a proportion, otherwise, an integer.}
}
\description{
Tools in epiworld
}
\examples{
epitool <- tool(
  name = "Vaccine",
  susceptibility_reduction = .9,
  transmission_reduction = .5,
  recovery_enhancer = .5, 
  death_reduction = .9
)
}
