% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus.R
\name{virus}
\alias{virus}
\alias{add_virus}
\alias{add_virus_n}
\alias{virus_set_status}
\title{Virus}
\usage{
virus(name, post_immunity, prob_infecting, prob_recovery, prob_death)

add_virus(m, v, prevalence)

add_virus_n(m, v, prevalence)

virus_set_status(v, init, end, removed)
}
\arguments{
\item{name}{of the virus}

\item{post_immunity, prob_infecting, prob_recovery, prob_death}{to be documented}
}
\description{
Virus
}
\examples{
mseirconn <- ModelSEIRCONN(
  name                = "COVID-19",
  prevalence          = 0.01, 
  n                   = 10000,
  reproductive_number = 4, 
  incubation_days     = 7, 
  prob_transmission   = 0.5,
  prob_recovery       = 0.99
)

delta <- virus("Delta Variant", 0, .5, .2, .01)
virus_set_status(delta, 1, 3, 3)

add_virus(mseirconn, delta, .3)
init(mseirconn, 100, 992)
mseirconn

run(mseirconn)
mseirconn
}
