% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{epiworld-data}
\alias{epiworld-data}
\alias{get_hist_total}
\alias{get_transition_probability}
\alias{get_reproductive_number}
\title{Accessing the database of \code{epiworld}}
\usage{
get_hist_total(x)

get_transition_probability(x)

get_reproductive_number(x)
}
\arguments{
\item{x}{An object of class \code{\link{epiworld_sir}}, \code{\link{epiworld_seir}}, etc.
any model.}
}
\description{
Accessing the database of \code{epiworld}
}
\examples{
# SEIR Connected
seirconn <- ModelSEIRCONN(
  "Disease",
  10000,
  .1,
  2, 
  .8,
  7,
  .3
)

set.seed(937)
run(seirconn, 50)

get_transition_probability(seirconn) 
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()}
}
\concept{Models}
