% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{epiworld-methods}
\alias{epiworld-methods}
\alias{init}
\alias{queuing_on}
\alias{queuing_off}
\alias{agents_smallworld}
\alias{run}
\alias{get_status}
\alias{get_param}
\alias{set_param}
\alias{set_name}
\alias{get_name}
\title{Common methods for predefined models of Epiworld}
\usage{
init(m, days, seed)

queuing_on(x)

queuing_off(x)

agents_smallworld(m, n, k, d, p)

run(m)

get_status(x)

get_param(x, pname)

set_param(x, pname, pval)

set_name(x, mname)

get_name(x)
}
\arguments{
\item{m}{Model object.}

\item{days}{Number of days.}

\item{seed}{Seed to set for initializing random number generator.}

\item{x}{An object of class \code{epiworld_model}.}

\item{n}{Number of individuals in the population.}

\item{k}{Number of ties in the small world network.}

\item{d}{Logical scalar. Whether the graph is directed or not.}

\item{p}{Probability of rewiring.}

\item{pname}{String, name of the parameter}

\item{pval}{Numeric. Value of the parameter}

\item{mname}{String. Name of the model}
}
\description{
Common methods for predefined models of Epiworld
}
