% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{epiworld-methods}
\alias{epiworld-methods}
\alias{init}
\alias{agents_smallworld}
\alias{run}
\title{Common methods for predefined models of Epiworld}
\usage{
init(m, days, seed)

agents_smallworld(m, n, k, d, p)

run(m)
}
\arguments{
\item{m}{Model object.}

\item{days}{Number of days.}

\item{seed}{Seed to set for initializing random number generator.}

\item{n}{Number of individuals in the population.}

\item{k}{Number of ties in the small world network.}

\item{d}{Logical scalar. Whether the graph is directed or not.}

\item{p}{Probability of rewiring.}
}
\description{
Common methods for predefined models of Epiworld
}
