% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{epiworld-methods}
\alias{epiworld-methods}
\alias{queuing_on}
\alias{queuing_off}
\alias{agents_smallworld}
\alias{run}
\alias{run_multiple}
\alias{get_state}
\alias{get_param}
\alias{set_param}
\alias{set_name}
\alias{get_name}
\title{Common methods for predefined models of Epiworld}
\usage{
queuing_on(x)

queuing_off(x)

agents_smallworld(m, n, k, d, p)

run(m, ndays, seed = sample.int(10000, 1))

run_multiple(
  m,
  ndays,
  nsims,
  seed = sample.int(10000, 1),
  saver = make_saver(),
  reset = TRUE,
  verbose = TRUE,
  nthreads = 1L
)

get_state(x)

get_param(x, pname)

set_param(x, pname, pval)

set_name(x, mname)

get_name(x)
}
\arguments{
\item{x}{An object of class \code{epiworld_model}.}

\item{m}{Model object.}

\item{n}{Number of individuals in the population.}

\item{k}{Number of ties in the small world network.}

\item{d}{Logical scalar. Whether the graph is directed or not.}

\item{p}{Probability of rewiring.}

\item{seed}{Seed to set for initializing random number generator.}

\item{pname}{String, name of the parameter}

\item{pval}{Numeric. Value of the parameter}

\item{mname}{String. Name of the model}

\item{days}{Number of days.}
}
\description{
Common methods for predefined models of Epiworld
}
