% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{epiworld-methods}
\alias{epiworld-methods}
\alias{queuing_on}
\alias{queuing_off}
\alias{verbose_off}
\alias{verbose_on}
\alias{run}
\alias{get_state}
\alias{get_param}
\alias{set_param}
\alias{set_name}
\alias{get_name}
\alias{get_n_variants}
\alias{get_n_tools}
\alias{get_ndays}
\alias{get_n_replicates}
\title{Common methods for predefined models of Epiworld}
\usage{
queuing_on(x)

queuing_off(x)

verbose_off(x)

verbose_on(x)

run(m, ndays, seed = sample.int(10000, 1))

get_state(x)

get_param(x, pname)

set_param(x, pname, pval)

set_name(x, mname)

get_name(x)

get_n_variants(x)

get_n_tools(x)

get_ndays(x)

get_n_replicates(x)
}
\arguments{
\item{x}{An object of class \code{epiworld_model}.}

\item{m}{Model object.}

\item{ndays}{Number of days (steps) of the simulation.}

\item{seed}{Seed to set for initializing random number generator.}

\item{pname}{String, name of the parameter}

\item{pval}{Numeric. Value of the parameter}

\item{mname}{String. Name of the model}
}
\description{
Common methods for predefined models of Epiworld
}
\details{
The \code{verbose_on} and \code{verbose_off} functions activate and deactivate printing
progress on screen, respectively. Both functions return the model (\code{x}) invisibly.
}
