% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{epiworld-methods}
\alias{epiworld-methods}
\alias{init}
\alias{agents_smallworld}
\alias{run}
\title{Common methods for predefined models of Epiworld}
\usage{
init(m, days, seed)

agents_smallworld(m, n, k, d, p)

run(m)
}
\arguments{
\item{m, days, seed, n, k, d, p}{to be explained}
}
\description{
Common methods for predefined models of Epiworld
}
