% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-renamed.R, R/global-actions.R
\name{epiworldR-deprecated}
\alias{epiworldR-deprecated}
\alias{globalaction_tool}
\alias{globalaction_tool_logit}
\alias{globalaction_set_params}
\alias{globalaction_fun}
\title{Deprecated functions in epiworldR}
\usage{
globalaction_tool(...)

globalaction_tool_logit(...)

globalaction_set_params(...)

globalaction_fun(...)
}
\arguments{
\item{...}{Arguments to be passed to the new function.}
}
\description{
Starting version 0.0-4, epiworld changed how it refered to "actions."
Following more traditional ABMs, actions are now called "events."
}
