% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R, R/functions-renamed.R
\name{globalaction_tool}
\alias{globalaction_tool}
\alias{globalaction_tool_logit}
\alias{globalaction_set_params}
\alias{globalaction_fun}
\alias{epiworldR-deprecated}
\alias{ModelMeaslesSchool}
\alias{ModelMeaslesMixing}
\alias{ModelMeaslesMixingRiskQuarantine}
\alias{ModelMeaslesQuarantine}
\alias{add_tool_n}
\alias{add_virus_n}
\title{Deprecated and removed functions in epiworldR}
\usage{
globalaction_tool(...)

globalaction_tool_logit(...)

globalaction_set_params(...)

globalaction_fun(...)

add_tool_n(model, tool, n)

add_virus_n(model, virus, n)
}
\arguments{
\item{...}{Arguments to be passed to the new function.}

\item{model}{Model object of class \code{epiworld_model}.}

\item{tool}{Tool object of class \code{epiworld_tool}.}

\item{n}{Deprecated.}

\item{virus}{Virus object of class \code{epiworld_virus}.}
}
\description{
Starting version 0.0-4, epiworld changed how it refered to "actions."
Following more traditional ABMs, actions are now called "events."

Starting version 0.11.0.0, the measles models have been removed from
\code{epiworldR} and now are part of the \code{measles} R package:
\url{https://github.com/UofUEpiBio/measles}. The models previously available
in \code{epiworldR} were: \code{ModelMeaslesSchool} and \code{ModelMeaslesMixing}.
}
\concept{deprecated-functions}
