\name{epiworldR-package}
\alias{epiworldR-package}
\alias{epiworldR}
\docType{package}
\title{
\packageTitle{epiworldR}
}
\description{
\packageDescription{epiworldR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{epiworldR}
\packageIndices{epiworldR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{epiworldR}

Maintainer: \packageMaintainer{epiworldR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
