% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiworldR-package.R
\docType{package}
\name{epiworldR-package}
\alias{epiworldR}
\alias{epiworldR-package}
\title{epiworldR}
\description{
A flexible framework for Agent-Based Models (ABM), the epiworldR package provides a fast method for prototyping disease outbreaks and transmission models. It supports multiple epidemiological models like the Susceptible-Infected-Susceptible (SIS,) Susceptible-Infected-Removed (SIR,) Susceptible-Exposed-Infected-Removed (SEIR,) and others, involving arbitrary mitigation policies and multiple-disease models. Users can specify infectiousness/susceptibility rates as a function of agents' features, providing great complexity for the model dynamics. epiworldR uses a C++ backend, which makes it very fast. Furthermore, epiworldR is ideal for simulation studies featuring large populations.
}
\author{
\strong{Maintainer}: Derek Meyer \email{derekmeyer37@gmail.com}

Authors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID})
}

Other contributors:
\itemize{
  \item Centers for Disease Control and Prevention (Award number 1U01CK000585; 75D30121F00003) [funder]
}

}
