% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFMCMC.R
\name{LFMCMC}
\alias{LFMCMC}
\alias{epiworld_lfmcmc}
\alias{run_lfmcmc}
\alias{set_observed_data}
\alias{set_proposal_fun}
\alias{set_simulation_fun}
\alias{set_summary_fun}
\alias{set_kernel_fun}
\alias{seed_lfmcmc}
\alias{set_par_names}
\alias{set_stats_names}
\alias{print.epiworld_lfmcmc}
\alias{use_proposal_norm_reflective}
\alias{use_kernel_fun_gaussian}
\title{Likelihood-Free Markhov Chain Monte Carlo (LFMCMC)}
\usage{
LFMCMC(model)

run_lfmcmc(lfmcmc, params_init_, n_samples_, epsilon_)

set_observed_data(lfmcmc, observed_data_)

set_proposal_fun(lfmcmc, fun)

set_simulation_fun(lfmcmc, fun)

set_summary_fun(lfmcmc, fun)

set_kernel_fun(lfmcmc, fun)

seed_lfmcmc(lfmcmc, s)

set_par_names(lfmcmc, names)

set_stats_names(lfmcmc, names)

\method{print}{epiworld_lfmcmc}(x, ...)

use_proposal_norm_reflective(lfmcmc)

use_kernel_fun_gaussian(lfmcmc)
}
\arguments{
\item{model}{A model of class \link{epiworld_model}}

\item{lfmcmc}{The LFMCMC model}

\item{params_init_}{Initial model parameters}

\item{n_samples_}{Number of samples}

\item{epsilon_}{Epsilon parameter}

\item{observed_data_}{Observed data}

\item{fun}{The LFMCMC kernel function}

\item{s}{The rand engine seed}

\item{names}{The model stats names}

\item{x}{LFMCMC model to print}

\item{...}{Ignored}
}
\value{
\itemize{
\item The \code{LFMCMC}function returns a model of class \link{epiworld_lfmcmc}.
}

The simulated model of class \code{epiworld_lfmcmc}.

The lfmcmc model with the observed data added

The lfmcmc model with the proposal function added

The lfmcmc model with the simulation function added

The lfmcmc model with the summary function added

The lfmcmc model with the kernel function added

The lfmcmc model with the seed set

The lfmcmc model with the parameter names added

The lfmcmc model with the stats names added

The lfmcmc model

The LFMCMC model with proposal function set to norm reflective

The LFMCMC model with kernel function set to gaussian
}
\description{
Likelihood-Free Markhov Chain Monte Carlo (LFMCMC)
}
\details{
TODO: Detail LFMCMC
}
\examples{
model_sir <- ModelSIR(name = "COVID-19", prevalence = 0.01,
  transmission_rate = 0.9, recovery_rate = 0.1)
model_lfmcmc <- LFMCMC(model_sir)
}
