% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_builder.R
\name{model_builder}
\alias{model_builder}
\alias{new_model}
\alias{add_state}
\alias{update_fun_susceptible}
\alias{epiworld_update_fun}
\alias{update_fun_rate}
\title{Model building functions}
\usage{
new_model()

add_state(model, state_name, update_fun = NULL)

update_fun_susceptible(exclude = integer(0L))

update_fun_rate(param_names, target_states)
}
\arguments{
\item{model}{An object of class \link{epiworld_model}.}

\item{state_name}{A string with the name of the state to be added.}

\item{update_fun}{An object of class \code{epiworld_update_fun} with
the update function to be used for the new state. A \code{NULL} value
can be used if the state does not update (e.g., a "dead" state).}
}
\value{
\itemize{
\item The function \code{new_model()} returns a new model object of class
\link{epiworld_model}.
}

\itemize{
\item The function \code{update_fun_susceptible()} returns an object of class
\code{epiworld_update_fun} that can be used as an update function for a
susceptible state.
}

\itemize{
\item The function \code{update_fun_rate()} returns an object of class
\code{epiworld_update_fun} that can be used as an update function
for a state that updates at a constant rate (e.g., recovery, death).
}
}
\description{
Functions to build models from scratch (or to modify existing models).
}
\examples{
# Create a new model
model <- new_model()

# Adding recovery rate
add_param(model, "Rec Rate", 0.1)
add_param(model, "Trans Rate", 0.5)

# Creating the update function for the susceptible state
update_fun_s <- update_fun_susceptible()

# Adding the susceptible state to the model
add_state(model, "S", update_fun_s)

# Creating the update function for the infected
update_fun_i <- update_fun_rate(
  param_names = "Rec Rate",
  target_states = 2L
)

# Adding the infected state to the model
add_state(model, "I", update_fun_i)

# Adding the recovered state to the model
add_state(model, "R", NULL)

# Creating a virus
flu <- virus(
  "Flu", 0, .5, .2, .01,
  prevalence = 0, as_proportion = TRUE
)

virus_set_state(flu, 1, 2, 2)

set_prob_infecting_ptr(flu, model, "Trans Rate")

set_distribution_virus(
  flu,
  distribute_virus_randomly(1L, FALSE)
)

# Adding the virus to the model
add_virus(model, flu)

# Creating a SBM network
agents_smallworld(model, n = 1000, k = 20, d = FALSE, p = .01)

# Running the model
run(model, ndays = 100, seed = 1912)

summary(model)

}
