% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelDiagram.R
\name{ModelDiagram}
\alias{ModelDiagram}
\alias{epiworld_modeldiagram}
\alias{draw_mermaid_from_data}
\alias{draw_mermaid_from_file}
\alias{draw_mermaid_from_files}
\title{Model Diagram}
\usage{
ModelDiagram()

draw_mermaid_from_data(
  model_diagram,
  states,
  transition_probs,
  output_file = "",
  allow_self_transitions = FALSE
)

draw_mermaid_from_file(
  model_diagram,
  transitions_file,
  output_file = "",
  allow_self_transitions = FALSE
)

draw_mermaid_from_files(
  model_diagram,
  transitions_files,
  output_file = "",
  allow_self_transitions = FALSE
)
}
\arguments{
\item{model_diagram}{ModelDiagram object}

\item{states}{String vector. List of model states.}

\item{transition_probs}{Numeric vector. Transition probability matrix}

\item{output_file}{String. Path to the output file.}

\item{allow_self_transitions}{Logical. Whether to allow self-transitions.}

\item{transitions_file}{String. Path to file containing the transition probabilities matrix.}

\item{transitions_files}{String vector. List of files containing transition probabilities matrices from multiple model runs.}
}
\value{
\itemize{
\item The \code{ModelDiagram} function returns an object of class \link{epiworld_modeldiagram}.
}

\itemize{
\item The ModelDiagram object.
}

\itemize{
\item The ModelDiagram object.
}

\itemize{
\item The ModelDiagram object.
}
}
\description{
Model Diagram
}
\details{
Draws diagrams based on transition probability matrices.
}
