% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIRCONN.R
\name{ModelSEIRCONN}
\alias{ModelSEIRCONN}
\alias{epiworld_seirconn}
\alias{plot.epiworld_seirconn}
\title{Susceptible Exposed Infected Removed model (SEIR connected)}
\usage{
ModelSEIRCONN(
  name,
  n,
  prevalence,
  contact_rate,
  prob_transmission,
  incubation_days,
  prob_recovery
)

\method{plot}{epiworld_seirconn}(x, main = "SEIRCONN Model", ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{n}{Number of individuals in the population.}

\item{prevalence}{Initial proportion of individuals with the virus.}

\item{contact_rate}{Numeric scalar. Average number of contacts per step.}

\item{prob_transmission}{Numeric scalar between 0 and 1. Probability of
transmission.}

\item{incubation_days}{Numeric scalar greater than 0. Average number of
incubation days.}

\item{prob_recovery}{Numeric scalar between 0 and 1. Probability of recovery.}

\item{x}{Object of class SEIRCONN.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\description{
The SEIR connected model implements a model where all agents are connected.
This is equivalent to a compartmental model.
}
\examples{
# An example with COVID-19
model_seirconn <- ModelSEIRCONN(
  name                = "COVID-19",
  prevalence          = 0.01, 
  n                   = 10000,
  contact_rate        = 2, 
  incubation_days     = 7, 
  prob_transmission   = 0.5,
  prob_recovery       = 0.3
)
  
# Running and printing
run(model_seirconn, ndays = 100, seed = 1912)
model_seirconn

plot(model_seirconn)

# Adding the flu
flu <- virus("Flu", .9, 1/7)
add_virus(model_seirconn, flu, .001)

#' # Running and printing
run(model_seirconn, ndays = 100, seed = 1912)
model_seirconn

plot(model_seirconn)
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
