% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSEIRCONN.R
\name{ModelSEIRCONN}
\alias{ModelSEIRCONN}
\alias{init.epiworld_seirconn}
\alias{print.epiworld_seirconn}
\alias{run.epiworld_seirconn}
\title{Susceptible Infected Susceptible model (SEIRCONN)}
\usage{
ModelSEIRCONN(
  name,
  n,
  prevalence,
  reproductive_number,
  prob_transmission,
  incubation_days,
  prob_recovery
)

\method{init}{epiworld_seirconn}(m, days, seed)

\method{print}{epiworld_seirconn}(x, ...)

\method{run}{epiworld_seirconn}(m)
}
\arguments{
\item{name}{Name of the virus}

\item{prevalence}{a number}

\item{reproductive_number}{a number}

\item{prob_transmission}{a number}

\item{incubation_days}{a number}

\item{prob_recovery}{a number}
}
\description{
Susceptible Infected Susceptible model (SEIRCONN)

#' @rdname ModelSEIRCONN
#' @export
agents_smallworld.epiworld_seirconn <- function(m, n, k, d, p) {
agents_smallworld_sir(m, n, k, d, p)
}
}
\seealso{
Other Models: 
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()}
}
\concept{Models}
