% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIR.R
\name{ModelSIR}
\alias{ModelSIR}
\alias{init.epiworld_sir}
\alias{print.epiworld_sir}
\alias{agents_smallworld.epiworld_sir}
\alias{run.epiworld_sir}
\title{SIR model}
\usage{
ModelSIR(name, prevalence, infectiousness, recovery)

\method{init}{epiworld_sir}(m, days, seed)

\method{print}{epiworld_sir}(x, ...)

\method{agents_smallworld}{epiworld_sir}(m, n, k, d, p)

\method{run}{epiworld_sir}(m)
}
\arguments{
\item{name}{Name of the virus}

\item{prevalence}{a number}

\item{infectiousness}{a number}

\item{recovery}{a number}
}
\description{
SIR model
}
