% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIR.R
\name{ModelSIR}
\alias{ModelSIR}
\alias{epiworld_sir}
\alias{plot.epiworld_sir}
\title{SIR model}
\usage{
ModelSIR(name, prevalence, infectiousness, recovery)

\method{plot}{epiworld_sir}(x, main = "SIR Model", ...)
}
\arguments{
\item{name}{String. Name of the virus.}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{infectiousness}{Numeric scalar between 0 and 1. Virus's rate of
infection.}

\item{recovery}{Numeric scalar between 0 and 1. Rate of recovery from virus.}

\item{x}{Object of class SIR.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\description{
SIR model
}
\examples{
model_sir <- ModelSIR(name = "COVID-19", prevalence = 0.01, 
                      infectiousness = 0.9, recovery = 0.1)

# Adding a small world population
agents_smallworld(
  model_sir,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Running and printing
run(model_sir, ndays = 100, seed = 1912)
model_sir

# Plotting
plot(model_sir)
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
