% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIRCONN.R
\name{ModelSIRCONN}
\alias{ModelSIRCONN}
\alias{epiworld_sirconn}
\alias{init.epiworld_sirconn}
\alias{print.epiworld_sirconn}
\alias{run.epiworld_sirconn}
\alias{plot.epiworld_sirconn}
\title{Susceptible Infected Removed model (SIR connected)}
\usage{
ModelSIRCONN(
  name,
  n,
  prevalence,
  reproductive_number,
  prob_transmission,
  prob_recovery
)

\method{init}{epiworld_sirconn}(m, days, seed)

\method{print}{epiworld_sirconn}(x, ...)

\method{run}{epiworld_sirconn}(m)

\method{plot}{epiworld_sirconn}(x, ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{n}{Number of individuals in the population.}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{reproductive_number}{Numeric scalar. Reproductive number.}

\item{prob_transmission}{Numeric scalar between 0 and 1. Probability of
transmission.}

\item{prob_recovery}{Numeric scalar between 0 and 1. Probability of recovery.}

\item{x}{Object of class SIRCONN.}

\item{...}{Currently ignore.}

\item{k}{Number of ties in the small world network.}

\item{d}{Logical scalar. Whether the graph is directed or not.}

\item{p}{Probability of rewiring.}
}
\description{
Susceptible Infected Removed model (SIR connected)
}
\examples{
model_sirconn <- ModelSIRCONN(
  name                = "COVID-19",
  n                   = 10000,
  prevalence          = 0.01,
  reproductive_number = 5,
  prob_transmission   = 0.4,
  prob_recovery       = 0.95
)
  
# Initializing
init(model_sirconn, days = 100, seed = 1912)
# Running and printing
run(model_sirconn)
model_sirconn

}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
