% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIRCONN.R
\name{ModelSIRCONN}
\alias{ModelSIRCONN}
\alias{epiworld_sirconn}
\alias{plot.epiworld_sirconn}
\title{Susceptible Infected Removed model (SIR connected)}
\usage{
ModelSIRCONN(
  name,
  n,
  prevalence,
  contact_rate,
  prob_transmission,
  prob_recovery
)

\method{plot}{epiworld_sirconn}(x, main = get_name(x), ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{n}{Number of individuals in the population.}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{contact_rate}{Numeric scalar. Average number of contacts per step.}

\item{prob_transmission}{Numeric scalar between 0 and 1. Probability of
transmission.}

\item{prob_recovery}{Numeric scalar between 0 and 1. Probability of recovery.}

\item{x}{Object of class SIRCONN.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\value{
\itemize{
\item The \code{ModelSIRCONN}function returns a model of class \link{epiworld_model}.
}

The \code{plot} function returns a plot of the SIRCONN model of class
\link{epiworld_model}.
}
\description{
Susceptible Infected Removed model (SIR connected)
}
\examples{
model_sirconn <- ModelSIRCONN(
  name                = "COVID-19",
  n                   = 10000,
  prevalence          = 0.01,
  contact_rate        = 5,
  prob_transmission   = 0.4,
  prob_recovery       = 0.95
)
  
# Running and printing
run(model_sirconn, ndays = 100, seed = 1912)
model_sirconn

plot(model_sirconn,  main = "SIRCONN Model")
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelDiffNet}()},
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
