% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIRLogit.R
\name{ModelSIRLogit}
\alias{ModelSIRLogit}
\title{SIR Logistic model}
\usage{
ModelSIRLogit(
  vname,
  data,
  coefs_infect,
  coefs_recover,
  coef_infect_cols,
  coef_recover_cols,
  prob_infection,
  prob_recovery,
  prevalence
)
}
\arguments{
\item{vname}{Name of the virus}

\item{data}{A numeric matrix with \code{n} rows.}

\item{coefs_infect}{Numeric vector. Coefficients associated to infect.}

\item{coefs_recover}{Numeric vector. Coefficients associated to recover.}

\item{coef_infect_cols}{Integer vector. Columns in the coeficient.}

\item{coef_recover_cols}{Integer vector. Columns in the coeficient.}

\item{prob_infection}{Numeric scalar. Baseline probability of infection.}

\item{prob_recovery}{Numeric scalar. Baseline probability of recovery.}

\item{prevalence}{Numeric scalar. Prevalence (initial state) in proportion.}
}
\description{
SIR Logistic model
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
