% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSIS.R
\name{ModelSIS}
\alias{ModelSIS}
\alias{epiworld_sis}
\alias{plot.epiworld_sis}
\title{Susceptible Infected Susceptible model (SIS)}
\usage{
ModelSIS(name, prevalence, infectiousness, recovery)

\method{plot}{epiworld_sis}(x, main = "SIS Model", ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{prevalence}{Double. Initial proportion of individuals with the virus.}

\item{infectiousness}{Numeric scalar between 0 and 1. Virus's rate of
infection}

\item{recovery}{Numeric scalar between 0 and 1. Rate of recovery from virus.}

\item{x}{Object of class SIS.}

\item{main}{Title of the plot}

\item{...}{Currently ignore.}
}
\description{
Susceptible Infected Susceptible model (SIS)
}
\examples{
model_sis <- ModelSIS(name = "COVID-19", prevalence = 0.01, 
                     infectiousness = 0.9, recovery = 0.1)

# Adding a small world population
agents_smallworld(
  model_sis,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Running and printing
run(model_sis, ndays = 100, seed = 1912)
model_sis
}
\seealso{
epiworld-methods

Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIRLogit}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSURV}()},
\code{\link{epiworld-data}}
}
\concept{Models}
