% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSURV.R
\name{ModelSURV}
\alias{ModelSURV}
\alias{epiworld_surv}
\alias{init.epiworld_surv}
\alias{print.epiworld_surv}
\alias{agents_smallworld.epiworld_surv}
\alias{run.epiworld_surv}
\alias{plot.epiworld_surv}
\title{Susceptible Infected Susceptible model (SURV)}
\usage{
ModelSURV(
  name,
  prevalence,
  efficacy_vax,
  latent_period,
  prob_symptoms,
  prop_vaccinated,
  prop_vax_redux_transm,
  infect_period,
  prop_vax_redux_infect,
  surveillance_prob,
  prob_transmission,
  prob_death,
  prob_noreinfect
)

\method{init}{epiworld_surv}(m, days, seed)

\method{print}{epiworld_surv}(x, ...)

\method{agents_smallworld}{epiworld_surv}(m, n, k, d, p)

\method{run}{epiworld_surv}(m)

\method{plot}{epiworld_surv}(x, ...)
}
\arguments{
\item{name}{String. Name of the virus}

\item{prevalence}{Initial proportion of individuals with the virus.}

\item{efficacy_vax}{Double. Efficacy of the vaccine
(1 - P(acquire the disease)).}

\item{latent_period}{Double. Shape parameter of a 'Gamma(latent_period, 1)'
distribution. This coincides with the expected number of latent days.}

\item{prob_symptoms}{Double. Probability of generating symptoms.}

\item{prop_vaccinated}{Double. Probability of vaccination. Coincides with
the initial prevalence of vaccinated individuals.}

\item{prop_vax_redux_transm}{Double. Factor by which the vaccine reduces
transmissibility.}

\item{infect_period}{Double. Shape parameter of a 'Gamma(infected_period, 1)'
distribution. This coincides with the expected number of infectious days.}

\item{prop_vax_redux_infect}{Double. Factor by which the vaccine reduces
the chances of becoming infected.}

\item{surveillance_prob}{Double. Probability of testing an agent.}

\item{prob_transmission}{Double. Raw transmission probability.}

\item{prob_death}{Double. Raw probability of death for symptomatic
individuals.}

\item{prob_noreinfect}{Double. Probability of no re-infection.}

\item{m, days, seed, x, ..., n, k, d, p}{to be explained}
}
\description{
Susceptible Infected Susceptible model (SURV)
}
\examples{
model_surv <- ModelSURV(name = "COVID-19", prevalence = 0.2, 
                       efficacy_vax = 0.6, latent_period = 4, 
                       infect_period = 5, prob_symptoms = 0.5, 
                       prop_vaccinated = 0.7, prop_vax_redux_transm = 0.8, 
                       prop_vax_redux_infect = 0.95, surveillance_prob = 0.1, 
                       prob_transmission = 0.2, prob_death = 0.001, 
                       prob_noreinfect = 0.5)

# Adding a small world population
agents_smallworld(
  model_surv,
  n = 1000,
  k = 5,
  d = FALSE,
  p = .01
  )
  
# Initializing
init(model_surv, days = 100, seed = 1912)
# Running and printing
run(model_surv)
model_surv 

}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{epiworld-data}}
}
\concept{Models}
