% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelSURV.R
\name{ModelSURV}
\alias{ModelSURV}
\alias{epiworld_surv}
\alias{init.epiworld_surv}
\alias{print.epiworld_surv}
\alias{agents_smallworld.epiworld_surv}
\alias{run.epiworld_surv}
\alias{plot.epiworld_surv}
\title{Susceptible Infected Susceptible model (SURV)}
\usage{
ModelSURV(
  name,
  prevalence,
  efficacy_vax,
  latent_period,
  prob_symptoms,
  prop_vaccinated,
  prop_vax_redux_transm,
  infect_period,
  prop_vax_redux_infect,
  surveillance_prob,
  prob_transmission,
  prob_death,
  prob_noreinfect
)

\method{init}{epiworld_surv}(m, days, seed)

\method{print}{epiworld_surv}(x, ...)

\method{agents_smallworld}{epiworld_surv}(m, n, k, d, p)

\method{run}{epiworld_surv}(m)

\method{plot}{epiworld_surv}(x, ...)
}
\arguments{
\item{name}{Name of the virus}

\item{prevalence}{a number}

\item{efficacy_vax}{to be explained}

\item{latent_period}{to be explained}

\item{prob_symptoms}{to be explained}

\item{prop_vaccinated}{to be explained}

\item{prop_vax_redux_transm}{to be explained}

\item{infect_period}{to be explained}

\item{prop_vax_redux_infect}{to be explained}

\item{surveillance_prob}{to be explained}

\item{prob_transmission}{to be explained}

\item{prob_death}{to be explained}

\item{prob_noreinfect}{to be explained}

\item{m, days, seed, x, ..., n, k, d, p}{to be explained}
}
\description{
Susceptible Infected Susceptible model (SURV)
}
\seealso{
Other Models: 
\code{\link{ModelSEIRCONN}()},
\code{\link{ModelSEIR}()},
\code{\link{ModelSIRCONN}()},
\code{\link{ModelSIR}()},
\code{\link{ModelSIS}()},
\code{\link{epiworld-data}}
}
\concept{Models}
