% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_saver.R
\name{run_multiple}
\alias{run_multiple}
\alias{read_multiple_runs}
\alias{make_saver}
\title{Run multiple simulations at once}
\usage{
run_multiple(
  m,
  ndays,
  nsims,
  seed = sample.int(10000, 1),
  saver = make_saver(),
  reset = TRUE,
  verbose = TRUE,
  nthreads = 1L
)

read_multiple_runs(saver)

make_saver(what = c("total_hist"), fn = "")
}
\arguments{
\item{m, ndays, seed}{See \link{run}.}

\item{nsims}{Integer. Number of replicats}

\item{saver}{An object of class \link{epiworld_saver}.}

\item{nthreads}{Currently ignored}

\item{what}{Character vector with the things to save (see details)}

\item{fn}{A file name pattern.}
}
\value{
In the case of \code{make_saver}, an list of class \code{epiworld_saver}.
The function \code{run_multiple} returns a list of the same class as \code{m}; usually
an \code{epiworld_modwel} object..
}
\description{
Run multiple simulations at once
}
\details{
Currently, the following elements can be saved:
\itemize{
\item \code{total_hist}
\item \code{variant_info}
\item \code{variant_hist}
\item \code{tool_info}
\item \code{tool_hist}
\item \code{transmission}
\item \code{transition}
\item \code{reproductive}
\item \code{generation}
}
}
