% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tool.R
\name{tool}
\alias{tool}
\alias{add_tool}
\alias{add_tool_n}
\title{Tools in epiworld}
\usage{
tool(
  name,
  susceptibility_reduction,
  transmission_reduction,
  recovery_enhancer,
  death_reduction
)

add_tool(m, t, prevalence)

add_tool_n(m, t, prevalence)
}
\arguments{
\item{m}{Model}
}
\description{
Tools in epiworld
}
\examples{
epitool <- tool(
  "Vaccine",
  .9,
  .3,
  .5,
  .9
)

}
