% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virus.R
\name{virus}
\alias{virus}
\alias{add_virus}
\alias{add_virus_n}
\alias{virus_set_state}
\alias{rm_virus}
\title{Virus}
\usage{
virus(
  name,
  prob_infecting,
  prob_recovery = 0.5,
  prob_death = 0,
  post_immunity = 1
)

add_virus(m, v, prevalence)

add_virus_n(m, v, prevalence)

virus_set_state(v, init, end, removed)

rm_virus(m, virus_pos)
}
\arguments{
\item{name}{of the virus}

\item{post_immunity, prob_infecting, prob_recovery, prob_death}{to be documented}

\item{m}{An object of class \code{epiworld-model}.}

\item{v}{An object of class \code{epiworld_virus}}

\item{prevalence}{In the case of \code{add_virus}, a proportion, otherwise, an integer.}

\item{init, end, removed}{states after acquiring a virus, removing a virus,
and removing the agent as a result of the virus, respectively.}

\item{virus_pos}{Positive integer. Index of the virus's position in the model.}
}
\description{
Virus
}
\examples{
mseirconn <- ModelSEIRCONN(
  name                = "COVID-19",
  prevalence          = 0.01, 
  n                   = 10000,
  contact_rate        = 4, 
  incubation_days     = 7, 
  prob_transmission   = 0.5,
  prob_recovery       = 0.99
)

delta <- virus("Delta Variant", 0, .5, .2, .01)
virus_set_state(delta, 1, 3, 3)

add_virus(mseirconn, delta, .3)

run(mseirconn, ndays = 100, seed = 992)
mseirconn
}
